closes #21; Agregada la funcionalidad de manipular coordenadas geográficas en las ubicaciones para los proyectos de API en SQLite y PostgreSQL

+++ CervezasColombia_CS_API_SQLite_Dapper/CervezasColombia_CS_API_SQLite_Dapper/Models/EnvasadoCerveza.cs
@@ -6,6 +6,6 @@
         public string Nombre { get; set; } = string.Empty;
         public int Unidad_Volumen_Id { get; set; } = 0;
         public string Unidad_Volumen { get; set; } = string.Empty;
-        public int Volumen { get; set; } = 0;
+        public float Volumen { get; set; } = 0.0f;
     }
 }
